Cast to correct type
